Broke UserService into UserService + UserServiceImpl to abide by Dependency Injection
