+++ src/main/java/com/example/musicAll/Controller/UserController.java
@@ -1,7 +1,7 @@
 package com.example.musicAll.Controller;
 
 import com.example.musicAll.Model.User;
-import com.example.musicAll.Service.UserService;
+import com.example.musicAll.Service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,7 +11,7 @@ import java.util.List;
 @RequestMapping("/user")
 public class UserController {
     @Autowired
-    UserService userService;
+    UserServiceImpl userService;
     /**
      * Creates /helloworld as an endpoint
      * @return Hello World!!!
