+++ src/main/java/com/example/musicAll/Service/UserService.java
@@ -1,27 +1,12 @@
 package com.example.musicAll.Service;
 
 import com.example.musicAll.Model.User;
-import com.example.musicAll.Repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-public class UserService {
-    @Autowired
-    private UserRepository userRepository;
+public interface UserService {
+    public List<User> listUsers();
+    public User createUser(User newUser);
+    public User login(String username, String password);
 
-    public List<User> listUsers() {
-        return userRepository.findAll();
-    }
-
-    public User createUser(User newUser){
-        return userRepository.save(newUser);
-    }
-
-    public User login(String username, String password) {
-        return userRepository.login(username, password);
-    }
 }
