+++ src/main/java/com/example/musicAll/Service/UserServiceImpl.java
@@ -1,0 +1,30 @@
+package com.example.musicAll.Service;
+
+import com.example.musicAll.Model.User;
+import com.example.musicAll.Repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class UserServiceImpl implements UserService {
+    @Autowired
+    private UserRepository userRepository;
+
+    @Override
+    public List<User> listUsers() {
+        return userRepository.findAll();
+    }
+
+    @Override
+    public User createUser(User newUser){
+        return userRepository.save(newUser);
+    }
+
+    @Override
+    public User login(String username, String password) {
+        return userRepository.login(username, password);
+    }
+}
